Get rid of busybox dockerhub image usage in tests
DockerHub is rate limiting, so at some point pulling this will simply fail.
